server/requirements.txt
flask>=2.3,<4.0
requests>=2.31,<3.0
python-dotenv>=1.0,<2.0
